myapp/requirements.txt
opencv-python==4.4.0.44
pycocotools==2.0.2
flask==1.1.2
jsonpickle==1.4.1
h5py==2.10.0
matplotlib
opencv-contrib-python
requests
tqdm
flask-wtf
nvidia-ml-py3
pretrainedmodels
efficientnet_pytorch
resnest
scikit-learn
scikit-image
imgaug
albumentations
cython
pycocotools
seaborn
thop
graphviz
pretrainedmodels==0.7.4
efficientnet-pytorch==0.6.3
timm==0.3.2
segmentation-models-pytorch
albumentations
